requirements.txt
asgiref==3.3.1
astroid==2.4.2
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.24.0
confusable-homoglyphs==3.2.0
Django==3.1.5
django-bootstrap4==2.3.1
django-decouple==2.1
django-registration==3.1.1
django-tinymce==3.2.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
pkg-resources==0.0.0
psycopg2==2.8.6
pylint==2.6.0
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.2
wrapt==1.12.1
